move harmonics cache to harmonics
